moving code to correct card and minor styling adjustments
